iox-#1391 Move 'system_configuration.hpp' to new location
